Remove EventListView and implementation plus EventListToolbar and implementation, move code to EventListFragment.
