--- eventdiscovery-sdk/src/main/java/com/schedjoules/eventdiscovery/eventlist/view/EventListView.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2016 SchedJoules
- *
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.schedjoules.eventdiscovery.eventlist.view;
-
-import android.support.v7.widget.RecyclerView;
-
-
-/**
- * Represents the Event list UI component, i.e. the {@link RecyclerView} and it's {@link RecyclerView.Adapter}.
- *
- * @author Gabor Keszthelyi
- */
-public interface EventListView<A extends RecyclerView.Adapter>
-{
-
-    /**
-     * The {@link RecyclerView.Adapter} of the {@link RecyclerView}.
-     */
-    A adapter();
-
-}
